requirements.txt
python3-smbus
mpu6050-raspberrypi
